+++ rust/operator-binary/src/main.rs
@@ -46,6 +46,10 @@ struct OpaRun {
     #[clap(long, env)]
     opa_bundle_builder_clusterrole: String,
 
+    /// The full image tag of the operator, used to deploy the user_info_fetcher.
+    #[clap(long, env)]
+    operator_image: String,
+
     #[clap(flatten)]
     common: ProductOperatorRun,
 }
@@ -55,10 +59,11 @@ async fn main() -> Result<(), error::Error> {
     let opts = Opts::parse();
     match opts.cmd {
         Command::Crd => {
-            OpaCluster::print_yaml_schema()?;
+            OpaCluster::print_yaml_schema(built_info::CARGO_PKG_VERSION)?;
         }
         Command::Run(OpaRun {
             opa_bundle_builder_clusterrole: opa_builder_clusterrole,
+            operator_image,
             common:
                 ProductOperatorRun {
                     product_config,
@@ -91,6 +96,7 @@ async fn main() -> Result<(), error::Error> {
                 product_config,
                 watch_namespace,
                 opa_builder_clusterrole,
+                operator_image,
             )
             .await?;
         }
@@ -107,6 +113,7 @@ async fn create_controller(
     product_config: ProductConfigManager,
     watch_namespace: WatchNamespace,
     opa_bundle_builder_clusterrole: String,
+    user_info_fetcher_image: String,
 ) -> OperatorResult<()> {
     let opa_api: Api<OpaCluster> = watch_namespace.get_api(&client);
     let daemonsets_api: Api<DaemonSet> = watch_namespace.get_api(&client);
@@ -126,6 +133,7 @@ async fn create_controller(
                 client: client.clone(),
                 product_config,
                 opa_bundle_builder_clusterrole,
+                user_info_fetcher_image,
             }),
         )
         .map(|res| {