+++ rust/user-info-fetcher/src/http_error.rs
@@ -1,0 +1,55 @@
+use std::sync::Arc;
+
+use axum::{response::IntoResponse, Json};
+use hyper::StatusCode;
+use serde::Serialize;
+
+pub trait Error: std::error::Error {
+    fn status_code(&self) -> StatusCode;
+}
+impl<T: Error> Error for Arc<T> {
+    fn status_code(&self) -> StatusCode {
+        let inner: &T = self;
+        inner.status_code()
+    }
+}
+
+pub struct JsonResponse<E> {
+    pub error: E,
+}
+
+impl<E> From<E> for JsonResponse<E> {
+    fn from(error: E) -> Self {
+        Self { error }
+    }
+}
+
+impl<E: Error> IntoResponse for JsonResponse<E> {
+    fn into_response(self) -> axum::response::Response {
+        (
+            self.error.status_code(),
+            Json(Container {
+                error: Payload {
+                    message: self.error.to_string(),
+                    causes: std::iter::successors(self.error.source(), |err| err.source())
+                        .map(|err| err.to_string())
+                        .collect(),
+                },
+            }),
+        )
+            .into_response()
+    }
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+struct Container {
+    error: Payload,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+struct Payload {
+    message: String,
+    causes: Vec<String>,
+}